X rotation
Playing around with how to align AC and PC with 0 slope between the two. Trying out angles between z and y to rotate the volume in x. Program taking forever to run the images... Makes it challenging to visually compare volumes to determine best rotation. Will work on that in upcoming days